move tree and tree-group renderer into own module; also extract the functions for the context menu and hoist where possible
