+++ frontend/javascripts/oxalis/view/right-border-tabs/tree_hierarchy_view.tsx
@@ -1,83 +1,49 @@
+import { DownOutlined } from "@ant-design/icons";
+import { Tree as AntdTree, GetRef, MenuProps, Modal, TreeProps } from "antd";
+import React, { useCallback, useEffect, useRef, useState } from "react";
 import { AutoSizer } from "react-virtualized";
-import { MenuProps, Modal, notification, Tree as AntdTree, GetRef, TreeProps } from "antd";
-import {
-  DeleteOutlined,
-  PlusOutlined,
-  ShrinkOutlined,
-  ExpandAltOutlined,
-  ArrowRightOutlined,
-  DownOutlined,
-  FolderOutlined,
-} from "@ant-design/icons";
-import { useDispatch } from "react-redux";
-import { batchActions } from "redux-batched-actions";
-import React, { useState, useEffect, useRef, useCallback } from "react";
-import _ from "lodash";
-import type { Action } from "oxalis/model/actions/actions";
-import {
-  TreeTypeEnum,
-  LongUnitToShortUnitMap,
-  type TreeType,
-  type Vector3,
-} from "oxalis/constants";
 
+import { mapGroups } from "oxalis/model/accessors/skeletontracing_accessor";
 import {
-  getGroupByIdWithSubgroups,
-  TreeNode,
-  MISSING_GROUP_ID,
-  callDeep,
-  createGroupToTreesMap,
-  insertTreesAndTransform,
-  makeBasicGroupObject,
-  anySatisfyDeep,
-  GroupTypeEnum,
-  deepFlatFilter,
-  getNodeKey,
-  moveGroupsHelper,
-  findParentGroupNode,
-} from "oxalis/view/right-border-tabs/tree_hierarchy_view_helpers";
-import type { TreeMap, TreeGroup } from "oxalis/store";
-import { getMaximumGroupId } from "oxalis/model/reducers/skeletontracing_reducer_helpers";
-import {
-  setActiveTreeAction,
-  setActiveTreeGroupAction,
-  setTreeColorAction,
+  setTreeGroupAction,
+  toggleAllTreesAction,
   toggleTreeAction,
   toggleTreeGroupAction,
-  toggleAllTreesAction,
-  setTreeGroupsAction,
-  shuffleTreeColorAction,
-  setTreeGroupAction,
-  deleteTreeAction,
-  toggleInactiveTreesAction,
-  shuffleAllTreeColorsAction,
-  setTreeEdgeVisibilityAction,
-  setTreeTypeAction,
 } from "oxalis/model/actions/skeletontracing_actions";
-import messages from "messages";
-import { formatNumberToLength, formatLengthAsVx } from "libs/format_utils";
-import { api, Store } from "oxalis/singletons";
-import { ChangeColorMenuItemContent } from "components/color_picker";
-import { HideTreeEdgesIcon } from "./hide_tree_eges_icon";
-import { ColoredDotIcon } from "./segments_tab/segment_list_item";
-import { mapGroups } from "oxalis/model/accessors/skeletontracing_accessor";
+import { Store } from "oxalis/singletons";
+import type { TreeGroup } from "oxalis/store";
+import {
+  createGroupToTreesMap,
+  deepFlatFilter,
+  findParentGroupNode,
+  getNodeKey,
+  GroupTypeEnum,
+  insertTreesAndTransform,
+  MISSING_GROUP_ID,
+  moveGroupsHelper,
+  TreeNode,
+} from "oxalis/view/right-border-tabs/tree_hierarchy_view_helpers";
 import { getContextMenuPositionFromEvent } from "../context_menu";
-import FastTooltip from "components/fast_tooltip";
 import { ContextMenuContainer } from "./sidebar_context_menu";
+import {
+  onBatchActions,
+  Props,
+  renderGroupNode,
+  renderTreeNode,
+  selectGroupById,
+  setUpdateTreeGroups,
+} from "./tree_hierarchy_renderers";
 
-type Props = {
-  activeTreeId: number | null | undefined;
-  activeGroupId: number | null | undefined;
-  treeGroups: TreeGroup[];
-  sortBy: string;
-  trees: TreeMap;
-  selectedTreeIds: number[];
-  onSingleSelectTree: (treeId: number, dispatchSetActiveTree: boolean) => void;
-  onMultiSelectTree: (treeId: number) => void;
-  onRangeSelectTrees: (treeIds: number[]) => void;
-  deselectAllTrees: () => void;
-  onDeleteGroup: (arg0: number) => void;
-  allowUpdate: boolean;
+const onCheck: TreeProps<TreeNode>["onCheck"] = (_checkedKeysValue, info) => {
+  const { id, type } = info.node;
+
+  if (type === GroupTypeEnum.TREE) {
+    Store.dispatch(toggleTreeAction(id));
+  } else if (id === MISSING_GROUP_ID) {
+    Store.dispatch(toggleAllTreesAction());
+  } else {
+    Store.dispatch(toggleTreeGroupAction(id));
+  }
 };
 
 function TreeHierarchyView(props: Props) {
@@ -87,32 +53,7 @@ function TreeHierarchyView(props: Props) {
   const [contextMenuPosition, setContextMenuPosition] = useState<[number, number] | null>(null);
   const [menu, setMenu] = useState<MenuProps | null>(null);
 
-  const onOpenContextMenu = (menu: MenuProps, event: React.MouseEvent<HTMLDivElement>) => {
-    event.preventDefault();
-
-    const [x, y] = getContextMenuPositionFromEvent(event, "tree-list-context-menu-overlay");
-    showContextMenuAt(x, y, menu);
-  };
-
-  const showContextMenuAt = useCallback((xPos: number, yPos: number, menu: MenuProps) => {
-    // On Windows the right click to open the context menu is also triggered for the overlay
-    // of the context menu. This causes the context menu to instantly close after opening.
-    // Therefore delay the state update to delay that the context menu is rendered.
-    // Thus the context overlay does not get the right click as an event and therefore does not close.
-    setTimeout(() => {
-      setContextMenuPosition([xPos, yPos]);
-      setMenu(menu);
-    }, 0);
-  }, []);
-
-  const hideContextMenu = useCallback(() => {
-    setContextMenuPosition(null);
-    setMenu(null);
-  }, []);
-
   const treeRef = useRef<GetRef<typeof AntdTree>>(null);
-
-  const dispatch = useDispatch();
 
   useEffect(() => {
     // equivalent of LifeCycle hook "getDerivedStateFromProps"
@@ -162,6 +103,29 @@ function TreeHierarchyView(props: Props) {
     }
   }, [props.activeGroupId]);
 
+  const onOpenContextMenu = (menu: MenuProps, event: React.MouseEvent<HTMLDivElement>) => {
+    event.preventDefault();
+
+    const [x, y] = getContextMenuPositionFromEvent(event, "tree-list-context-menu-overlay");
+    showContextMenuAt(x, y, menu);
+  };
+
+  const showContextMenuAt = useCallback((xPos: number, yPos: number, menu: MenuProps) => {
+    // On Windows the right click to open the context menu is also triggered for the overlay
+    // of the context menu. This causes the context menu to instantly close after opening.
+    // Therefore delay the state update to delay that the context menu is rendered.
+    // Thus the context overlay does not get the right click as an event and therefore does not close.
+    setTimeout(() => {
+      setContextMenuPosition([xPos, yPos]);
+      setMenu(menu);
+    }, 0);
+  }, []);
+
+  const hideContextMenu = useCallback(() => {
+    setContextMenuPosition(null);
+    setMenu(null);
+  }, []);
+
   const onExpand: TreeProps<TreeNode>["onExpand"] = (expandedKeys, info) => {
     const clickedNode = info.node;
     const expandedKeySet = new Set(expandedKeys);
@@ -183,18 +147,6 @@ function TreeHierarchyView(props: Props) {
       }
     });
     setUpdateTreeGroups(newGroups);
-  };
-
-  const onCheck: TreeProps<TreeNode>["onCheck"] = (_checkedKeysValue, info) => {
-    const { id, type } = info.node;
-
-    if (type === GroupTypeEnum.TREE) {
-      toggleTree(id);
-    } else if (id === MISSING_GROUP_ID) {
-      setToggleAllTrees();
-    } else {
-      toggleTreeGroup(id);
-    }
   };
 
   function onSelectTreeNode(node: TreeNode, evt: MouseEvent) {
@@ -237,11 +189,6 @@ function TreeHierarchyView(props: Props) {
     }
   }
 
-  function selectGroupById(groupId: number) {
-    props.deselectAllTrees();
-    setActiveTreeGroup(groupId);
-  }
-
   function onSelectGroupNode(node: TreeNode) {
     const groupId = node.id;
     const numberOfSelectedTrees = props.selectedTreeIds.length;
@@ -252,62 +199,13 @@ function TreeHierarchyView(props: Props) {
         content: `You have ${numberOfSelectedTrees} selected Trees. Do you really want to select this group?
         This will deselect all selected trees.`,
         onOk: () => {
-          selectGroupById(groupId);
+          selectGroupById(props.deselectAllTrees, groupId);
         },
 
         onCancel() {},
       });
     } else {
-      selectGroupById(groupId);
-    }
-  }
-
-  function setExpansionOfAllSubgroupsTo(parentGroup: TreeNode, expanded: boolean) {
-    if (parentGroup.id === MISSING_GROUP_ID) {
-      const newGroups = mapGroups(props.treeGroups, (group) => {
-        if (group.isExpanded !== expanded) {
-          return { ...group, isExpanded: expanded };
-        }
-        return group;
-      });
-      setUpdateTreeGroups(newGroups);
-      return;
-    }
-    const subGroups = getGroupByIdWithSubgroups(props.treeGroups, parentGroup.id);
-    const subGroupsMap = new Set(subGroups);
-    // If the subgroups should be collapsed, do not collapse the group itself.
-    // Do expand the group if the subgroups are expanded though.
-    if (expanded === false) subGroupsMap.delete(parentGroup.id);
-    const newGroups = mapGroups(props.treeGroups, (group) => {
-      if (subGroupsMap.has(group.groupId) && expanded !== group.isExpanded) {
-        return { ...group, isExpanded: expanded };
-      } else {
-        return group;
-      }
-    });
-    setUpdateTreeGroups(newGroups);
-  }
-
-  function onMoveWithContextAction(targetParentNode: TreeNode) {
-    const activeComponent = getLabelForActiveItems();
-    const targetGroupId = targetParentNode.id === MISSING_GROUP_ID ? null : targetParentNode.id;
-    let allTreesToMove;
-    if (activeComponent === "tree") {
-      allTreesToMove = [props.activeTreeId];
-    } else if (activeComponent === "trees") {
-      allTreesToMove = props.selectedTreeIds;
-    }
-    if (allTreesToMove) {
-      const moveActions = allTreesToMove.map((treeId) =>
-        setTreeGroupAction(
-          targetGroupId,
-          // @ts-expect-error ts-migrate(2345) FIXME: Argument of type 'number' is not assignable to par... Remove this comment to see the full error message
-          treeId,
-        ),
-      );
-      onBatchActions(moveActions, "SET_TREE_GROUP");
-    } else if (activeComponent === "group" && props.activeGroupId != null) {
-      api.tracing.moveSkeletonGroup(props.activeGroupId, targetGroupId);
+      selectGroupById(props.deselectAllTrees, groupId);
     }
   }
 
@@ -349,284 +247,8 @@ function TreeHierarchyView(props: Props) {
     setUpdateTreeGroups(newGroups);
   }
 
-  function createGroup(groupId: number) {
-    const newTreeGroups = _.cloneDeep(props.treeGroups);
-
-    const newGroupId = getMaximumGroupId(newTreeGroups) + 1;
-    const newGroup = makeBasicGroupObject(newGroupId, `Group ${newGroupId}`);
-
-    if (groupId === MISSING_GROUP_ID) {
-      newTreeGroups.push(newGroup);
-    } else {
-      callDeep(newTreeGroups, groupId, (item) => {
-        item.children.push(newGroup);
-      });
-    }
-
-    setUpdateTreeGroups(newTreeGroups);
-    selectGroupById(newGroupId);
-  }
-
-  function shuffleTreeGroupColors(groupId: number) {
-    const groupToTreeMap = createGroupToTreesMap(props.trees);
-    const groupIdWithSubgroups = getGroupByIdWithSubgroups(props.treeGroups, groupId);
-    const shuffleTreeColorActions = groupIdWithSubgroups.flatMap((subGroupId) => {
-      if (subGroupId in groupToTreeMap)
-        return groupToTreeMap[subGroupId].map((tree) => shuffleTreeColorAction(tree.treeId));
-      return [];
-    });
-    onBatchActions(shuffleTreeColorActions, "SHUFFLE_TREE_COLOR");
-  }
-
-  function setTreeGroupColor(groupId: number, color: Vector3) {
-    const groupToTreeMap = createGroupToTreesMap(props.trees);
-    const groupIdWithSubgroups = getGroupByIdWithSubgroups(props.treeGroups, groupId);
-    const setTreeColorActions = groupIdWithSubgroups.flatMap((subGroupId) => {
-      if (subGroupId in groupToTreeMap)
-        return groupToTreeMap[subGroupId].map((tree) => setTreeColorAction(tree.treeId, color));
-      return [];
-    });
-    onBatchActions(setTreeColorActions, "SET_TREE_COLOR");
-  }
-
-  function setAllTreesColor(color: Vector3) {
-    const setTreeColorActions = Object.values(props.trees).map((tree) =>
-      setTreeColorAction(tree.treeId, color),
-    );
-    onBatchActions(setTreeColorActions, "SET_TREE_COLOR");
-  }
-
-  function renderGroupNode(node: TreeNode) {
-    // The root group must not be removed or renamed
-    const { id, name } = node;
-
-    const isEditingDisabled = !props.allowUpdate;
-    const hasSubgroup = anySatisfyDeep(
-      node.children,
-      (child) => child.type === GroupTypeEnum.GROUP,
-    );
-    const labelForActiveItems = getLabelForActiveItems();
-    const createMenu = (): MenuProps => ({
-      items: [
-        {
-          key: "create",
-          onClick: () => {
-            createGroup(id);
-            hideContextMenu();
-          },
-          disabled: isEditingDisabled,
-          icon: <PlusOutlined />,
-          label: "Create new group",
-        },
-        labelForActiveItems != null
-          ? {
-              key: "moveHere",
-              onClick: () => {
-                onMoveWithContextAction(node);
-                hideContextMenu();
-              },
-              disabled: isEditingDisabled,
-              icon: <ArrowRightOutlined />,
-              label: `Move active ${labelForActiveItems} here`,
-            }
-          : null,
-        {
-          key: "delete",
-          disabled: isEditingDisabled,
-          onClick: () => props.onDeleteGroup(id),
-          icon: <DeleteOutlined />,
-          label: "Delete group",
-        },
-        hasSubgroup
-          ? {
-              key: "collapseSubgroups",
-              onClick: () => {
-                setExpansionOfAllSubgroupsTo(node, false);
-                hideContextMenu();
-              },
-              icon: <ShrinkOutlined />,
-              label: "Collapse all subgroups",
-            }
-          : null,
-        hasSubgroup
-          ? {
-              key: "expandSubgroups",
-              onClick: () => {
-                setExpansionOfAllSubgroupsTo(node, true);
-                hideContextMenu();
-              },
-              icon: <ExpandAltOutlined />,
-              label: "Expand all subgroups",
-            }
-          : null,
-        {
-          key: "hideTree",
-          onClick: () => {
-            setActiveTreeGroup(id);
-            toggleHideInactiveTrees();
-            hideContextMenu();
-          },
-          icon: <i className="fas fa-eye" />,
-          label: "Hide/Show all other trees",
-        },
-        {
-          key: "shuffleTreeGroupColors",
-          onClick: () => {
-            if (id === MISSING_GROUP_ID) shuffleAllTreeColors();
-            else shuffleTreeGroupColors(id);
-          },
-          icon: <i className="fas fa-adjust" />,
-          label: "Shuffle Tree Group Colors",
-        },
-        {
-          key: "setTreeGroupColor",
-          disabled: isEditingDisabled,
-          icon: <i className="fas fa-eye-dropper fa-sm " />,
-          label: (
-            <ChangeColorMenuItemContent
-              title="Change Tree Group Color"
-              isDisabled={isEditingDisabled}
-              onSetColor={(color) => {
-                if (id === MISSING_GROUP_ID) setAllTreesColor(color);
-                else setTreeGroupColor(id, color);
-              }}
-              rgb={[0.5, 0.5, 0.5]}
-            />
-          ),
-        },
-      ],
-    });
-
-    // Make sure the displayed name is not empty
-    const displayableName = name.trim() || "<Unnamed Group>";
-    return (
-      <div className="nowrap" onContextMenu={(evt) => onOpenContextMenu(createMenu(), evt)}>
-        <span>
-          <FolderOutlined className="icon-margin-right" />
-          {displayableName}
-        </span>
-      </div>
-    );
-  }
-
-  function renderTreeNode(node: TreeNode): React.ReactNode {
-    const tree = props.trees[node.id];
-    if (!tree) return null;
-
-    const isEditingDisabled = !props.allowUpdate;
-    const isAgglomerateSkeleton = tree.type === TreeTypeEnum.AGGLOMERATE;
-
-    const createMenu = (): MenuProps => {
-      return {
-        items: [
-          {
-            key: "changeTreeColor",
-            disabled: isEditingDisabled,
-            icon: <i className="fas fa-eye-dropper fa-sm " />,
-            label: (
-              <ChangeColorMenuItemContent
-                title="Change Tree Color"
-                isDisabled={isEditingDisabled}
-                onSetColor={(color) => {
-                  setTreeColor(tree.treeId, color);
-                }}
-                rgb={tree.color}
-              />
-            ),
-          },
-          {
-            key: "shuffleTreeColor",
-            onClick: () => shuffleTreeColor(tree.treeId),
-            title: "Shuffle Tree Color",
-            disabled: isEditingDisabled,
-            icon: <i className="fas fa-adjust" />,
-            label: "Shuffle Tree Color",
-          },
-          {
-            key: "deleteTree",
-            onClick: () => deleteTree(tree.treeId),
-            title: "Delete Tree",
-            disabled: isEditingDisabled,
-            icon: <i className="fas fa-trash" />,
-            label: "Delete Tree",
-          },
-          {
-            key: "measureSkeleton",
-            onClick: () => {
-              handleMeasureSkeletonLength(tree.treeId, tree.name);
-              hideContextMenu();
-            },
-            title: "Measure Tree Length",
-            icon: <i className="fas fa-ruler" />,
-            label: "Measure Tree Length",
-          },
-          {
-            key: "hideTree",
-            onClick: () => {
-              setActiveTree(tree.treeId);
-              toggleHideInactiveTrees();
-              hideContextMenu();
-            },
-            title: "Hide/Show All Other Trees",
-            icon: <i className="fas fa-eye" />,
-            label: "Hide/Show All Other Trees",
-          },
-          {
-            key: "hideTreeEdges",
-            onClick: () => {
-              setActiveTree(tree.treeId);
-              setTreeEdgesVisibility(tree.treeId, !tree.edgesAreVisible);
-              hideContextMenu();
-            },
-            title: "Hide/Show Edges of This Tree",
-            icon: <HideTreeEdgesIcon />,
-            label: "Hide/Show Edges of This Tree",
-          },
-          isAgglomerateSkeleton
-            ? {
-                key: "convertToNormalSkeleton",
-                onClick: () => {
-                  setTreeType(tree.treeId, TreeTypeEnum.DEFAULT);
-                  hideContextMenu();
-                },
-                title: "Convert to Normal Tree",
-                icon: <span className="fas fa-clipboard-check" />,
-                label: "Convert to Normal Tree",
-              }
-            : null,
-        ],
-      };
-    };
-
-    const maybeProofreadingIcon =
-      tree.type === TreeTypeEnum.AGGLOMERATE ? (
-        <FastTooltip title="Agglomerate Skeleton">
-          <i className="fas fa-clipboard-check icon-margin-right" />
-        </FastTooltip>
-      ) : null;
-
-    return (
-      <div className="nowrap" onContextMenu={(evt) => onOpenContextMenu(createMenu(), evt)}>
-        <ColoredDotIcon colorRGBA={[...tree.color, 1.0]} />
-        {`(${tree.nodes.size()}) `} {maybeProofreadingIcon} {tree.name}
-      </div>
-    );
-  }
-
   function isNodeDraggable(node: TreeNode): boolean {
     return props.allowUpdate && node.id !== MISSING_GROUP_ID;
-  }
-
-  function getLabelForActiveItems(): "trees" | "tree" | "group" | null {
-    // Only one type of component can be selected. It is not possible to select multiple groups.
-    if (props.selectedTreeIds.length > 0) {
-      return "trees";
-    } else if (props.activeTreeId != null) {
-      return "tree";
-    } else if (props.activeGroupId != null) {
-      return "group";
-    }
-    return null;
   }
 
   // checkedKeys includes all nodes with a "selected" checkbox
@@ -660,7 +282,9 @@ function TreeHierarchyView(props: Props) {
               height={height}
               ref={treeRef}
               titleRender={(node) =>
-                node.type === GroupTypeEnum.TREE ? renderTreeNode(node) : renderGroupNode(node)
+                node.type === GroupTypeEnum.TREE
+                  ? renderTreeNode(props, onOpenContextMenu, hideContextMenu, node)
+                  : renderGroupNode(props, onOpenContextMenu, hideContextMenu, node)
               }
               switcherIcon={<DownOutlined />}
               onSelect={(_selectedKeys, info: { node: TreeNode; nativeEvent: MouseEvent }) =>
@@ -689,77 +313,6 @@ function TreeHierarchyView(props: Props) {
       </AutoSizer>
     </>
   );
-
-  function deleteTree(treeId: number) {
-    props.deselectAllTrees();
-    dispatch(deleteTreeAction(treeId));
-  }
-}
-
-function handleMeasureSkeletonLength(treeId: number, treeName: string) {
-  const dataSourceUnit = Store.getState().dataset.dataSource.scale.unit;
-  const [lengthInUnit, lengthInVx] = api.tracing.measureTreeLength(treeId);
-
-  notification.open({
-    message: messages["tracing.tree_length_notification"](
-      treeName,
-      formatNumberToLength(lengthInUnit, LongUnitToShortUnitMap[dataSourceUnit]),
-      formatLengthAsVx(lengthInVx),
-    ),
-    icon: <i className="fas fa-ruler" />,
-  });
-}
-
-function setActiveTree(treeId: number) {
-  Store.dispatch(setActiveTreeAction(treeId));
-}
-
-function setActiveTreeGroup(groupId: number) {
-  Store.dispatch(setActiveTreeGroupAction(groupId));
-}
-
-function setTreeColor(treeId: number, color: Vector3) {
-  Store.dispatch(setTreeColorAction(treeId, color));
-}
-
-function shuffleTreeColor(treeId: number) {
-  Store.dispatch(shuffleTreeColorAction(treeId));
-}
-
-function toggleTree(treeId: number) {
-  Store.dispatch(toggleTreeAction(treeId));
-}
-
-function setTreeEdgesVisibility(treeId: number, edgesAreVisible: boolean) {
-  Store.dispatch(setTreeEdgeVisibilityAction(treeId, edgesAreVisible));
-}
-
-function toggleTreeGroup(groupId: number) {
-  Store.dispatch(toggleTreeGroupAction(groupId));
-}
-
-function setToggleAllTrees() {
-  Store.dispatch(toggleAllTreesAction());
-}
-
-function setUpdateTreeGroups(treeGroups: TreeGroup[]) {
-  Store.dispatch(setTreeGroupsAction(treeGroups));
-}
-
-function onBatchActions(actions: Action[], actionName: string) {
-  Store.dispatch(batchActions(actions, actionName));
-}
-
-function toggleHideInactiveTrees() {
-  Store.dispatch(toggleInactiveTreesAction());
-}
-
-function shuffleAllTreeColors() {
-  Store.dispatch(shuffleAllTreeColorsAction());
-}
-
-function setTreeType(treeId: number, type: TreeType) {
-  Store.dispatch(setTreeTypeAction(treeId, type));
 }
 
 // React.memo is used to prevent the component from re-rendering without the props changing
